[bpu][exu][ifu][util] remove deprecated constructs

+++ src/main/scala/exu/issue.scala
@@ -31,7 +31,7 @@ trait IssueUnitConstants
    // invalid  : slot holds no valid uop.
    // s_valid_1: slot holds a valid uop.
    // s_valid_2: slot holds a store-like uop that may be broken into two micro-ops.
-   val s_invalid :: s_valid_1 :: s_valid_2 :: Nil = Enum(UInt(), 3)
+   val s_invalid :: s_valid_1 :: s_valid_2 :: Nil = Enum(3)
 }
 
 // What physical register is broadcasting its wakeup?
